[BOJ] 폴더 이름 변경 char_string > five_char_string
