--- hansaem/boj/step/char_string/input/p05622.txt
@@ -1,2 +1,0 @@
-WA
-UNUCIC
