+++ hansaem/boj/step/five_char_string/b5_p27866.java
@@ -1,25 +1,23 @@
-package boj.step.char_string;
+package boj.step.five_char_string;
 
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class b2_02908 {
+public class b5_p27866 {
     public static void main(String[] args) throws IOException {
         //BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedReader br = new BufferedReader(
-                new FileReader("hansaem/boj/step/char_string/input/p02908.txt")
+                new FileReader("hansaem/boj/step/char_string/input/p27866.txt")
         );
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
         StringBuilder sb = new StringBuilder();
 
-        StringBuilder sb1 = new StringBuilder(st.nextToken());
-        StringBuilder sb2 = new StringBuilder(st.nextToken());
+        String str = String.valueOf(st.nextToken());
+        st = new StringTokenizer(br.readLine());
+        int num = Integer.parseInt(st.nextToken());
 
-        int num1 = Integer.parseInt(sb1.reverse().toString());
-        int num2 = Integer.parseInt(sb2.reverse().toString());
-
-        System.out.println(Math.max(num1, num2));
+        System.out.println(str.charAt(num - 1));
     }
 }
