+++ hansaem/boj/step/five_char_string/input/p01152.txt
