+++ hansaem/boj/step/five_char_string/input/p02675.txt
