+++ hansaem/boj/step/five_char_string/input/p05622.txt
@@ -1,0 +1,1 @@
+pulljima
