+++ hansaem/boj/step/five_char_string/input/p27866.txt
